make some addecuation of sales

+++ src/main/java/org/barbershop/sale/application/SaleCommand.java
@@ -1,13 +1,14 @@
 package org.barbershop.sale.application;
 
 import java.util.List;
+import java.math.BigDecimal;
 import org.barbershop.sale.domain.PaymentMethod;
 
 public record SaleCommand(
     Long customerId,
     Long employeeId,
     PaymentMethod paymentMethod,
-    Double discount,
+    BigDecimal discount,
     String notes,
     List<SaleItemCommand> items
 ) {
